+++ packages/install-expo-modules/src/plugins/android/withAndroidModulesMainActivity.ts
@@ -19,6 +19,11 @@ export const withAndroidModulesMainActivity: ConfigPlugin = config => {
 export function setModulesMainActivity(mainActivity: string, language: 'java' | 'kt'): string {
   const isJava = language === 'java';
 
+  if (mainActivity.match(/\s+ReactActivityDelegateWrapper\(/m) != null) {
+    // Early return if `ReactActivityDelegateWrapper` is already added.
+    return mainActivity;
+  }
+
   if (mainActivity.match(/\s+createReactActivityDelegate\(\)/m) == null) {
     // If not override `createReactActivityDelegate()`, tries to override with wrapper
     mainActivity = addImports(
@@ -49,7 +54,36 @@ export function setModulesMainActivity(mainActivity: string, language: 'java' | 
       'class MainActivity',
       addReactActivityDelegateBlock.join('\n')
     );
-  } else if (mainActivity.match(/\s+ReactActivityDelegateWrapper\(/m) == null) {
+  } else if (
+    // java: public static class MainActivityDelegate extends ReactActivityDelegate {
+    mainActivity.match(/\s+MainActivityDelegate\s+extends\s+ReactActivityDelegate\s+\{/) != null ||
+    // kotlin: class MainActivityDelegate(activity: ReactActivity?, mainComponentName: String?) : ReactActivityDelegate
+    mainActivity.match(/\s+MainActivityDelegate\(.+\)\s+:\s+ReactActivityDelegate.+\{/) != null
+  ) {
+    // If override `createReactActivityDelegate()` already, wrap it with `ReactActivityDelegateWrapper` for react-native 0.68+
+    mainActivity = addImports(mainActivity, ['expo.modules.ReactActivityDelegateWrapper'], isJava);
+
+    const newInstanceCodeBlock = findNewInstanceCodeBlock(
+      mainActivity,
+      'MainActivityDelegate',
+      language
+    );
+    if (newInstanceCodeBlock == null) {
+      throw new Error('Unable to find MainActivityDelegate new instance code block.');
+    }
+
+    const replacement = isJava
+      ? `new ReactActivityDelegateWrapper(this, ${newInstanceCodeBlock.code})`
+      : `ReactActivityDelegateWrapper(this, ${newInstanceCodeBlock.code})`;
+    mainActivity = replaceContentsWithOffset(
+      mainActivity,
+      replacement,
+      newInstanceCodeBlock.start,
+      newInstanceCodeBlock.end
+    );
+
+    return mainActivity;
+  } else {
     // If override `createReactActivityDelegate()` already, wrap it with `ReactActivityDelegateWrapper`
     mainActivity = addImports(mainActivity, ['expo.modules.ReactActivityDelegateWrapper'], isJava);